pyproject.toml
[project]
name = "context-forge"
version = "0.1.0"
description = "高性能动态上下文组装引擎 — LLM 应用的 ORM"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "Context Forge Contributors" },
]
keywords = ["llm", "context-engineering", "prompt", "rag", "token-budget"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# [Design Decision] 全部作为直接依赖，降低用户安装摩擦。
# 框架集成（langchain/llamaindex/haystack）使用 lazy import，
# 因为它们的依赖树庞大且可能互相冲突。
dependencies = [
    # --- 数据模型与校验 ---
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",

    # --- Token 计数 ---
    "tiktoken>=0.7.0",

    # --- CLI 与终端输出 ---
    "typer>=0.12.0",
    "rich>=13.0",

    # --- HTTP 服务 ---
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",

    # --- 可观测性 ---
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",

    # --- 缓存 ---
    "redis>=5.0",

    # --- 语义计算（语义缓存 / 去重） ---
    "sentence-transformers>=3.0",

    # --- 序列化 ---
    "protobuf>=5.0",

    # --- 异步支持 ---
    "anyio>=4.0",
]

[project.optional-dependencies]
# 框架集成作为可选依赖（依赖树过大，避免冲突）
langchain = ["langchain-core>=0.2"]
llamaindex = ["llama-index-core>=0.10"]
haystack = ["haystack-ai>=2.0"]

dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "pre-commit>=3.7",
    "httpx>=0.27",          # FastAPI 测试客户端
]

docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[project.urls]
Homepage = "https://github.com/context-forge/context-forge"
Documentation = "https://context-forge.github.io"
Repository = "https://github.com/context-forge/context-forge"
Issues = "https://github.com/context-forge/context-forge/issues"

# [DX Decision] CLI 入口点命名为 context-forge，与包名一致，便于记忆
[project.scripts]
context-forge = "context_forge.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = ["NUL", "*/NUL"]

[tool.hatch.build.targets.wheel]
packages = ["src/context_forge"]

# --- 开发工具配置 ---

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff 特有规则
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=context_forge --cov-report=term-missing"

[tool.coverage.run]
source = ["src/context_forge"]
branch = true
omit = [
    "tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
fail_under = 85
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
